// Frontend/src/app/weather1/Models/TemperatureData.ts
export class TemperatureData {
    temperature: number = 0; // ou null
    summaryImage: string = '';
    day: string = '';
    time: string = '';
    summaryPhrase: string = '';
    rainPercent: number = 0;
  }
  


/*export class TemperatureData{
    temperature:number;
    summaryImage:string;
    day:string;
    time:string;
    summaryPhrase:string;
    rainPercent:number;
    //location:string;
}*/