// Frontend/src/app/weather1/Models/TodaysHighlight.ts
export class TodaysHighlight {
    uvIndex: number = 0;
    windStatus: number = 0;
    sunrise: string = '';
    sunset: string = '';
    humidity: number = 0;
    visibility: number = 0;
    airQuality: number = 0;
  }
  

/*
export class TodaysHighlight{
    uvIndex:number;
    windStatus:number;
    sunrise: string;
    sunset: string;
    humidity:number;
    visibility:number;
    airQuality:number;
}
*/